refactored some code from the init, added option to upload the model
